requirements.txt
amqp==2.3.2
ansible==2.6.3
appnope==0.1.0
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
backcall==0.1.0
bcrypt==3.1.4
billiard==3.5.0.4
cachetools==3.1.1
celery==4.2.0
certifi==2019.9.11
cffi==1.11.5
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
daphne==1.4.2
decorator==4.3.0
Django==2.1.5
django-bootstrap3==11.0.0
django-celery-beat==1.3.0
django-celery-results==1.0.4
django-pure-pagination==0.3.0
django-redis==4.10.0
django-timezone-field==3.0
djangorestframework==3.9.0
gitdb2==2.0.5
GitPython==3.0.2
google-auth==1.6.3
harbor-py-light==2.1.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
Jinja2==2.10
kombu==4.2.1
kubernetes==10.0.1
MarkupSafe==1.1.0
msgpack-python==0.5.6
multi-key-dict==2.0.3
numpy==1.15.4
oauthlib==3.1.0
paramiko==2.4.2
parso==0.3.1
pbr==5.4.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
Pygments==2.3.0
PyHamcrest==1.9.0
PyMySQL==0.9.2
PyNaCl==1.3.0
python-crontab==2.3.5
python-dateutil==2.7.5
python-jenkins==1.5.0
python-magic==0.4.3
pytz==2018.7
PyYAML==4.2b1
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
six==1.11.0
smmap2==2.0.5
sqlparse==0.3.0
tqdm==4.28.1
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.25.6
vine==1.1.4
wcwidth==0.1.7
websocket-client==0.56.0
zope.interface==4.6.0
